HOLO-203 <refactor: Funnel jsx 로직 컴포넌트로 분리>
